+++ CHANGES.md
@@ -4,6 +4,8 @@ Change Log
 #### next release (8.1.22)
 
 * Added play story button in mobile view when there is an active story
+* `IDEAL ZOOM` can be customised by providing `idealZoom` property in `MappableTraits`.
+* [The next improvement]
 
 #### 8.1.21